net-im/spectrum2: drop "xmpp" USE flag
"xmpp" USE flag is pointless, as it bears no conditional dependencies
and doesn't make much difference in build time. It is also expected to
be used by most of users.

net-im/swift is an unconditional dependency and it appears to be more
than a mistake in the upstream build scripts, so won't get fixed.

Spectrum's parts *legitimately* depending on net-im/swift:
* libtransport (core)
* XMPP backend
* XMPP frontend
* Slack frontend
(There are some more parts which have an incorrect dependency on swiften
due to build script mistakes.)

Bug: https://bugs.gentoo.org/671798
Package-Manager: Portage-2.3.49, Repoman-2.3.12
Signed-off-by: Andrey Utkin <andrey_utkin@gentoo.org>